feat: updated hover card, added buttons, styled the card & the buttons, fixed show/hide issues & other minor fixes

+++ src/pages/content/unsubscribeHoverCard.ts
@@ -7,23 +7,24 @@ export interface IHoverCardElements {
 
 const handleUnsubscribe = (ev: MouseEvent) => {
   ev.stopPropagation();
-  const { email } = (window as any).mailPurge;
+  const { email } = mailPurgeGlobalVariables;
   console.log('clicked: unsubscribeButton', email);
 };
 
 // handle mouseover on hoverCard
-const handleMouseOverHoverCard = () => {
-  (window as any).mailPurge = { ...(window as any).mailPurge, isMouseOverHoverCard: true };
+const handleMouseOverHoverCard = (ev: MouseEvent) => {
+  ev.stopPropagation();
+  mailPurgeGlobalVariables.isMouseOverHoverCard = true;
 };
 // handle mouseout on hoverCard
 const handleMouseOutHoverCard = () => {
-  (window as any).mailPurge = { ...(window as any).mailPurge, isMouseOverHoverCard: false };
-  const isMouseOverMailPurgeBtn = (window as any).mailPurge.isMouseOverMailPurgeBtn || false;
+  const { hoverCardElements, mainBtnContainerId } = mailPurgeGlobalVariables;
 
-  if(!isMouseOverMailPurgeBtn){
-    hideHoverCard({parentEl, hoverCardElements})
-  }
+  setTimeout(() => {
+    hideHoverCard({ parentElId: mainBtnContainerId, hoverCardElements });
+  }, 500);
 
+  mailPurgeGlobalVariables.isMouseOverHoverCard = false;
 };
 
 const initializeHoverCard = (): IHoverCardElements => {
@@ -31,6 +32,7 @@ const initializeHoverCard = (): IHoverCardElements => {
   // main container
   const hoverCard = document.createElement('div');
   const label = document.createElement('p');
+  const btnContainer = document.createElement('div');
   const unsubscribeBtn = document.createElement('button');
   const deleteAllMailsBtn = document.createElement('button');
   const unsubscribeAndDeleteAllMailsBtn = document.createElement('button');
@@ -38,47 +40,63 @@ const initializeHoverCard = (): IHoverCardElements => {
   /// add classnames
   hoverCard.classList.add('hoverCard');
   label.classList.add('hoverCard-label');
+  btnContainer.classList.add('hoverCard-btnContainer');
   unsubscribeBtn.classList.add('hoverCard-unsubscribeBtn');
   deleteAllMailsBtn.classList.add('hoverCard-deleteAllMailsBtn');
   unsubscribeAndDeleteAllMailsBtn.classList.add('hoverCard-unsubscribeAndDeleteAllMailsBtn');
 
+  // add text to buttons
+  unsubscribeBtn.innerHTML = 'Unsubscribe';
+  deleteAllMailsBtn.innerHTML = 'Delete All Mails';
+  unsubscribeAndDeleteAllMailsBtn.innerHTML = 'Unsubscribe + Delete All Mails';
+
   // add text to label
   label.innerText = 'MailPurge: Email Actions for {name} & {email}';
 
+  // append buttons to the btnContainer
+  btnContainer.append(unsubscribeBtn, deleteAllMailsBtn, unsubscribeAndDeleteAllMailsBtn);
+
   // append  elements to parent el (Card)
-  hoverCard.append(label, unsubscribeBtn, deleteAllMailsBtn, unsubscribeAndDeleteAllMailsBtn);
+  hoverCard.append(label, btnContainer);
 
   return { hoverCard, unsubscribeBtn, deleteAllMailsBtn, unsubscribeAndDeleteAllMailsBtn };
 };
 // hide hoverCard
 
 type ShowHoverCardParams = {
-  parentEl: HTMLDivElement;
+  parentElId: string;
   hoverCardElements: IHoverCardElements;
   email: string;
   name: string;
 };
-const showHoverCard = ({ parentEl, hoverCardElements, email, name }: ShowHoverCardParams) => {
+
+const showHoverCard = ({ parentElId, hoverCardElements, email, name }: ShowHoverCardParams) => {
   const { hoverCard, unsubscribeBtn, deleteAllMailsBtn, unsubscribeAndDeleteAllMailsBtn } = hoverCardElements;
-  // hoverCard.style.top = `${parentEl.offsetTop}px`;
+
+  // get parent el from id
+  const parentEl = document.getElementById(parentElId);
+
   parentEl.appendChild(hoverCard);
 
-  (window as any).mailPurge = { email, name };
+  mailPurgeGlobalVariables.email = email;
+  mailPurgeGlobalVariables.name = name;
+
+  // stop event propagation for card container
+  hoverCard.addEventListener('click', ev => {
+    ev.stopPropagation();
+  });
 
   // add mouseover (on hover) event to card container
-
   hoverCard.addEventListener('mouseover', handleMouseOverHoverCard);
+  // add mouseout (on hover) event to card container
   hoverCard.addEventListener('mouseout', handleMouseOutHoverCard);
 
   // add onClick listener to buttons
-  // unsubscribe btn
   unsubscribeBtn.addEventListener('click', handleUnsubscribe);
-  // deleteAllMails btn
   deleteAllMailsBtn.addEventListener('click', (ev: MouseEvent) => {
     ev.stopPropagation();
     console.log('clicked: deleteAllMails', email, name);
   });
-  // unsubscribeAndDeleteAllMails btn
   unsubscribeAndDeleteAllMailsBtn.addEventListener('click', (ev: MouseEvent) => {
     ev.stopPropagation();
     console.log('clicked: unsubscribeAndDeleteAllMails', email, name);
@@ -89,20 +107,27 @@ const showHoverCard = ({ parentEl, hoverCardElements, email, name }: ShowHoverCa
 
 // hide hoverCard
 type HideHoverCardParams = {
-  parentEl: Element;
+  parentElId: string;
   hoverCardElements: IHoverCardElements;
 };
 
-const hideHoverCard = ({ parentEl, hoverCardElements }: HideHoverCardParams) => {
+const hideHoverCard = ({ parentElId, hoverCardElements }: HideHoverCardParams) => {
   const { hoverCard, unsubscribeBtn, deleteAllMailsBtn, unsubscribeAndDeleteAllMailsBtn } = hoverCardElements;
 
-  hoverCard.style.display = 'none';
-  hoverCard.style.visibility = 'hidden';
+  if (mailPurgeGlobalVariables.isMouseOverHoverCard || mailPurgeGlobalVariables.isMouseOverMailPurgeBtn)
+    return;
+  const parentEl = document.getElementById(parentElId);
 
-  parentEl.removeChild(hoverCard);
-  unsubscribeBtn.removeEventListener('click', handleUnsubscribe);
-  deleteAllMailsBtn.removeEventListener('click', handleUnsubscribe);
-  unsubscribeAndDeleteAllMailsBtn.removeEventListener('click', handleUnsubscribe);
+  if (parentEl && parentEl.contains(hoverCard)) {
+    hoverCard.style.display = 'none';
+    hoverCard.style.visibility = 'hidden';
+    // get parent el from id
+
+    parentEl.removeChild(hoverCard);
+    unsubscribeBtn.removeEventListener('click', handleUnsubscribe);
+    deleteAllMailsBtn.removeEventListener('click', handleUnsubscribe);
+    unsubscribeAndDeleteAllMailsBtn.removeEventListener('click', handleUnsubscribe);
+  }
 };
 
 export { initializeHoverCard, showHoverCard, hideHoverCard };
